requirements.txt
Flask==2.2
SQLAlchemy

python-dotenv==0.19.2
Werkzeug==2.2.2
pytest==7.0.1
pytest-flask==1.2.0

requests==2.26.0

flask-sqlalchemy==3.0.2
flask-migrate==3.1.0
psycopg2-binary==2.9.1
alembic==1.7.4